fix: Resolved failing sample code.
