+++ src/Blazing.Mvvm.Sample.Wasm/ViewModels/TestKeyedNavigationViewModel.cs
@@ -23,19 +23,3 @@ public sealed class TestKeyedNavigationViewModel(IMvvmNavigationManager mvvmNavi
         MvvmNavigationManager.NavigateTo(key, @params);
     }
 }
-
-public interface ITestKeyedNavigationViewModel : IViewModelBase
-{
-    RelayCommand<string> TestNavigateCommand { get; }
-
-    /// <inheritdoc cref="TestNavigationBaseViewModel._queryString"/>
-    string? QueryString { get; set; }
-
-    /// <inheritdoc cref="TestNavigationBaseViewModel._test"/>
-    string? Test { get; set; }
-
-    // populated by MvvmComponentBase
-    string? Echo { get; set; }
-
-    RelayCommand HexTranslateNavigateCommand { get; }
-}
