+++ src/Blazing.Mvvm.Sample.WebApp/Blazing.Mvvm.Sample.WebApp.Client/Pages/TestKeyedNavigation.razor
@@ -2,7 +2,7 @@
 @page "/keyedtest/{echo}"
 
 @attribute [ViewModelKey(nameof(TestKeyedNavigationViewModel))]
-@inherits MvvmComponentBase<TestKeyedNavigationViewModel>
+@inherits MvvmComponentBase<ITestKeyedNavigationViewModel>
 
 <h3>Test Keyed Navigation</h3>
